Move `dummy_blob` into `test_math.py`
Notes:

- `dummy_blob` is used *only* in `test_math.py`; its
  implementation details are specifically tied to those tests.
